changes en model filter scripts and headmat_plot.
