+++ phase1/results/metrics_results/filtered_results/dt_Q2_best_models.csv
@@ -1,5 +1,2 @@
 ,PK model
-0,DT5
-1,DT6
-2,DT7
-3,DT8
+0,DT6
