+++ phase1/results/metrics_results/filtered_results/lrg_Q2_best_models.csv
@@ -41,7 +41,4 @@
 39,LRG40
 40,LRG61
 41,LRG62
-42,LRG63
-43,LRG64
-44,LRG65
-45,LRG66
+42,LRG64
