+++ phase1/results/metrics_results/filtered_results/rf_Q2_best_models.csv
@@ -20,24 +20,3 @@
 18,RF22
 19,RF23
 20,RF24
-21,RF49
-22,RF50
-23,RF51
-24,RF52
-25,RF53
-26,RF54
-27,RF55
-28,RF56
-29,RF57
-30,RF58
-31,RF59
-32,RF60
-33,RF66
-34,RF68
-35,RF69
-36,RF70
-37,RF71
-38,RF72
-39,RF73
-40,RF79
-41,RF82
