+++ phase1/results/metrics_results/filtered_results/svm_Q2_best_models.csv
@@ -5,27 +5,15 @@
 3,SVM6
 4,SVM7
 5,SVM8
-6,SVM9
-7,SVM10
-8,SVM13
-9,SVM14
-10,SVM15
-11,SVM16
-12,SVM17
-13,SVM18
-14,SVM21
-15,SVM22
-16,SVM25
-17,SVM26
-18,SVM29
-19,SVM30
-20,SVM31
-21,SVM32
-22,SVM35
-23,SVM36
-24,SVM37
-25,SVM38
-26,SVM43
-27,SVM44
-28,SVM45
-29,SVM46
+6,SVM10
+7,SVM13
+8,SVM14
+9,SVM15
+10,SVM16
+11,SVM17
+12,SVM18
+13,SVM21
+14,SVM22
+15,SVM26
+16,SVM29
+17,SVM30
