--- phase1/support_functions/headmap_vars.py
@@ -1,32 +1,0 @@
-bounds = [
-    0.50,
-    0.59,
-    0.70,
-    0.74,
-    0.76,
-    0.81,
-    0.85,
-    0.89,
-    0.93,
-    0.95,
-    0.97,
-    0.98,
-    1.0,
-]
-colors = [
-    "yellow",
-    "gold",
-    "orange",
-    "peru",
-    "pink",
-    "deeppink",
-    "mediumvioletred",
-    "blueviolet",
-    # "indigo",
-    "yellowgreen",
-    "forestgreen",
-    # "cyan",
-    "deepskyblue",
-    "dodgerblue",
-    "darkblue",
-]
